Added Correct Copy Constructor
Added a Copy constructor to each item. Previously items were returning a copy in an incorrect way. They now use a copy constructor and return a copy of itself in a more correct way. Also changed the visibility of some of the variables in the Item super class to private and added getters accordingly.