Added Some Skill Implementation
Added Implementation for some of the skills I wanted to add. The skillData class was created to store stats for skills like level description and cost. this makes tracking the skills easier. The actual implantations of the skills is in the farm class as avoid any chance of circular dependency.